--- a/ARMeilleure/Instructions/SoftFallback.cs
+++ b/ARMeilleure/Instructions/SoftFallback.cs
@@ -1017,15 +1017,15 @@ namespace ARMeilleure.Instructions
         {
             for (int e = 0; e <= 3; e++)
             {
-                uint t = ShaChoose(hash_abcd.GetUInt32(1),
-                                   hash_abcd.GetUInt32(2),
-                                   hash_abcd.GetUInt32(3));
+                uint t = ShaChoose(hash_abcd.Get<uint>(1),
+                                   hash_abcd.Get<uint>(2),
+                                   hash_abcd.Get<uint>(3));
 
-                hash_e += Rol(hash_abcd.GetUInt32(0), 5) + t + wk.GetUInt32(e);
+                hash_e += Rol(hash_abcd.Get<uint>(0), 5) + t + wk.Get<uint>(e);
 
-                t = Rol(hash_abcd.GetUInt32(1), 30);
+                t = Rol(hash_abcd.Get<uint>(1), 30);
 
-                hash_abcd.Insert(1, t);
+                hash_abcd.Set(1, t);
 
                 Rol32_160(ref hash_e, ref hash_abcd);
             }
@@ -1042,15 +1042,15 @@ namespace ARMeilleure.Instructions
         {
             for (int e = 0; e <= 3; e++)
             {
-                uint t = ShaMajority(hash_abcd.GetUInt32(1),
-                                     hash_abcd.GetUInt32(2),
-                                     hash_abcd.GetUInt32(3));
+                uint t = ShaMajority(hash_abcd.Get<uint>(1),
+                                     hash_abcd.Get<uint>(2),
+                                     hash_abcd.Get<uint>(3));
 
-                hash_e += Rol(hash_abcd.GetUInt32(0), 5) + t + wk.GetUInt32(e);
+                hash_e += Rol(hash_abcd.Get<uint>(0), 5) + t + wk.Get<uint>(e);
 
-                t = Rol(hash_abcd.GetUInt32(1), 30);
+                t = Rol(hash_abcd.Get<uint>(1), 30);
 
-                hash_abcd.Insert(1, t);
+                hash_abcd.Set(1, t);
 
                 Rol32_160(ref hash_e, ref hash_abcd);
             }
@@ -1062,15 +1062,15 @@ namespace ARMeilleure.Instructions
         {
             for (int e = 0; e <= 3; e++)
             {
-                uint t = ShaParity(hash_abcd.GetUInt32(1),
-                                   hash_abcd.GetUInt32(2),
-                                   hash_abcd.GetUInt32(3));
+                uint t = ShaParity(hash_abcd.Get<uint>(1),
+                                   hash_abcd.Get<uint>(2),
+                                   hash_abcd.Get<uint>(3));
 
-                hash_e += Rol(hash_abcd.GetUInt32(0), 5) + t + wk.GetUInt32(e);
+                hash_e += Rol(hash_abcd.Get<uint>(0), 5) + t + wk.Get<uint>(e);
 
-                t = Rol(hash_abcd.GetUInt32(1), 30);
+                t = Rol(hash_abcd.Get<uint>(1), 30);
 
-                hash_abcd.Insert(1, t);
+                hash_abcd.Set(1, t);
 
                 Rol32_160(ref hash_e, ref hash_abcd);
             }
@@ -1080,8 +1080,8 @@ namespace ARMeilleure.Instructions
 
         public static V128 Sha1SchedulePart1(V128 w0_3, V128 w4_7, V128 w8_11)
         {
-            ulong t2 = w4_7.GetUInt64(0);
-            ulong t1 = w0_3.GetUInt64(1);
+            ulong t2 = w4_7.Get<ulong>(0);
+            ulong t1 = w0_3.Get<ulong>(1);
 
             V128 result = new V128(t1, t2);
 
@@ -1092,20 +1092,20 @@ namespace ARMeilleure.Instructions
         {
             V128 t = tw0_3 ^ (w12_15 >> 32);
 
-            uint tE0 = t.GetUInt32(0);
-            uint tE1 = t.GetUInt32(1);
-            uint tE2 = t.GetUInt32(2);
-            uint tE3 = t.GetUInt32(3);
+            uint tE0 = t.Get<uint>(0);
+            uint tE1 = t.Get<uint>(1);
+            uint tE2 = t.Get<uint>(2);
+            uint tE3 = t.Get<uint>(3);
 
             return new V128(tE0.Rol(1), tE1.Rol(1), tE2.Rol(1), tE3.Rol(1) ^ tE0.Rol(2));
         }
 
         private static void Rol32_160(ref uint y, ref V128 x)
         {
-            uint xE3 = x.GetUInt32(3);
+            uint xE3 = x.Get<uint>(3);
 
             x <<= 32;
-            x.Insert(0, y);
+            x.Set(0, y);
 
             y = xE3;
         }
@@ -1148,13 +1148,13 @@ namespace ARMeilleure.Instructions
 
             for (int e = 0; e <= 3; e++)
             {
-                uint elt = (e <= 2 ? w0_3 : w4_7).GetUInt32(e <= 2 ? e + 1 : 0);
+                uint elt = (e <= 2 ? w0_3 : w4_7).Get<uint>(e <= 2 ? e + 1 : 0);
 
                 elt = elt.Ror(7) ^ elt.Ror(18) ^ elt.Lsr(3);
 
-                elt += w0_3.GetUInt32(e);
+                elt += w0_3.Get<uint>(e);
 
-                result.Insert(e, elt);
+                result.Set(e, elt);
             }
 
             return result;
@@ -1164,7 +1164,7 @@ namespace ARMeilleure.Instructions
         {
             V128 result = new V128();
 
-            ulong t1 = w12_15.GetUInt64(1);
+            ulong t1 = w12_15.Get<ulong>(1);
 
             for (int e = 0; e <= 1; e++)
             {
@@ -1172,12 +1172,12 @@ namespace ARMeilleure.Instructions
 
                 elt = elt.Ror(17) ^ elt.Ror(19) ^ elt.Lsr(10);
 
-                elt += w0_3.GetUInt32(e) + w8_11.GetUInt32(e + 1);
+                elt += w0_3.Get<uint>(e) + w8_11.Get<uint>(e + 1);
 
-                result.Insert(e, elt);
+                result.Set(e, elt);
             }
 
-            t1 = result.GetUInt64(0);
+            t1 = result.Get<ulong>(0);
 
             for (int e = 2; e <= 3; e++)
             {
@@ -1185,9 +1185,9 @@ namespace ARMeilleure.Instructions
 
                 elt = elt.Ror(17) ^ elt.Ror(19) ^ elt.Lsr(10);
 
-                elt += w0_3.GetUInt32(e) + (e == 2 ? w8_11 : w12_15).GetUInt32(e == 2 ? 3 : 0);
+                elt += w0_3.Get<uint>(e) + (e == 2 ? w8_11 : w12_15).Get<uint>(e == 2 ? 3 : 0);
 
-                result.Insert(e, elt);
+                result.Set(e, elt);
             }
 
             return result;
@@ -1197,23 +1197,23 @@ namespace ARMeilleure.Instructions
         {
             for (int e = 0; e <= 3; e++)
             {
-                uint chs = ShaChoose(y.GetUInt32(0),
-                                     y.GetUInt32(1),
-                                     y.GetUInt32(2));
+                uint chs = ShaChoose(y.Get<uint>(0),
+                                     y.Get<uint>(1),
+                                     y.Get<uint>(2));
 
-                uint maj = ShaMajority(x.GetUInt32(0),
-                                       x.GetUInt32(1),
-                                       x.GetUInt32(2));
+                uint maj = ShaMajority(x.Get<uint>(0),
+                                       x.Get<uint>(1),
+                                       x.Get<uint>(2));
 
-                uint t1 = y.GetUInt32(3) + ShaHashSigma1(y.GetUInt32(0)) + chs + w.GetUInt32(e);
+                uint t1 = y.Get<uint>(3) + ShaHashSigma1(y.Get<uint>(0)) + chs + w.Get<uint>(e);
 
-                uint t2 = t1 + x.GetUInt32(3);
+                uint t2 = t1 + x.Get<uint>(3);
 
-                x.Insert(3, t2);
+                x.Set(3, t2);
 
-                t2 = t1 + ShaHashSigma0(x.GetUInt32(0)) + maj;
+                t2 = t1 + ShaHashSigma0(x.Get<uint>(0)) + maj;
 
-                y.Insert(3, t2);
+                y.Set(3, t2);
 
                 Rol32_256(ref y, ref x);
             }
@@ -1223,14 +1223,14 @@ namespace ARMeilleure.Instructions
 
         private static void Rol32_256(ref V128 y, ref V128 x)
         {
-            uint yE3 = y.GetUInt32(3);
-            uint xE3 = x.GetUInt32(3);
+            uint yE3 = y.Get<uint>(3);
+            uint xE3 = x.Get<uint>(3);
 
             y <<= 32;
             x <<= 32;
 
-            y.Insert(0, xE3);
-            x.Insert(0, yE3);
+            y.Set(0, xE3);
+            x.Set(0, yE3);
         }
 
         private static uint ShaHashSigma0(uint x)