… Ryujinx.Tests/Cpu/CpuTest.cs
… Ryujinx.Tests/Cpu/CpuTest32.cs
… Ryujinx.Tests/Cpu/CpuTestMisc.cs