@FICTURE7 2020-04-08
For comparison, here is the change in codegen for `GetUInt32`

Before:
```asm
G_M31163_IG01:
       56                   push     rsi
       4883EC20             sub      rsp, 32
						;; bbWeight=1    PerfScore 1.25
G_M31163_IG02:
       83FA03               cmp      edx, 3
       7744                 ja       SHORT G_M31163_IG11
       8BC2                 mov      eax, edx
       488D1575000000       lea      rdx, [reloc @RWD00]
       8B1482               mov      edx, dword ptr [rdx+4*rax]
       4C8D05E8FFFFFF       lea      r8, G_M31163_IG02
       4903D0               add      rdx, r8
       FFE2                 jmp      rdx
						;; bbWeight=1    PerfScore 7.25
G_M31163_IG03:
       8B01                 mov      eax, dword ptr [rcx]
						;; bbWeight=0.50 PerfScore 1.00
G_M31163_IG04:
       4883C420             add      rsp, 32
       5E                   pop      rsi
       C3                   ret      
						;; bbWeight=0.50 PerfScore 0.88
G_M31163_IG05:
       488B01               mov      rax, qword ptr [rcx]
       48C1E820             shr      rax, 32
						;; bbWeight=0.50 PerfScore 2.00
G_M31163_IG06:
       4883C420             add      rsp, 32
       5E                   pop      rsi
       C3                   ret      
						;; bbWeight=0.50 PerfScore 0.88
G_M31163_IG07:
       8B4108               mov      eax, dword ptr [rcx+8]
						;; bbWeight=0.50 PerfScore 1.00
G_M31163_IG08:
       4883C420             add      rsp, 32
       5E                   pop      rsi
       C3                   ret      
						;; bbWeight=0.50 PerfScore 0.88
G_M31163_IG09:
       488B4108             mov      rax, qword ptr [rcx+8]
       48C1E820             shr      rax, 32
						;; bbWeight=0.50 PerfScore 2.00
G_M31163_IG10:
       4883C420             add      rsp, 32
       5E                   pop      rsi
       C3                   ret      
						;; bbWeight=0.50 PerfScore 0.88
G_M31163_IG11:
       48B9105C9B38FC7F0000 mov      rcx, 0x7FFC389B5C10
       E833B8B95F           call     CORINFO_HELP_NEWSFAST
       488BF0               mov      rsi, rax
       B901000000           mov      ecx, 1
       48BA30D79938FC7F0000 mov      rdx, 0x7FFC3899D730
       E8BC91CB5F           call     CORINFO_HELP_STRCNS
       488BD0               mov      rdx, rax
       488BCE               mov      rcx, rsi
       E8D958FFFF           call     ArgumentOutOfRangeException:.ctor(String):this
       488BCE               mov      rcx, rsi
       E8E949B55F           call     CORINFO_HELP_THROW
       CC                   int3     
						;; bbWeight=0    PerfScore 0.00
RWD00  dd	0000001Dh ; case G_M31163_IG03
       dd	00000025h ; case G_M31163_IG05
       dd	00000032h ; case G_M31163_IG07
       dd	0000003Bh ; case G_M31163_IG09


; Total bytes of code 136, prolog size 5, PerfScore 31.60, (MethodHash=bdca8644) for method V128:GetUInt32(int):int:this
; ============================================================
```

After:
```asm
G_M31166_IG01:
       4883EC28             sub      rsp, 40
						;; bbWeight=1    PerfScore 0.25
G_M31166_IG02:
       83FA03               cmp      edx, 3
       770B                 ja       SHORT G_M31166_IG05
						;; bbWeight=1    PerfScore 1.25
G_M31166_IG03:
       4863C2               movsxd   rax, edx
       8B0481               mov      eax, dword ptr [rcx+4*rax]
						;; bbWeight=1    PerfScore 2.25
G_M31166_IG04:
       4883C428             add      rsp, 40
       C3                   ret      
						;; bbWeight=1    PerfScore 1.25
G_M31166_IG05:
       E8178EFFFF           call     V128:ThrowIndexOutOfRange()
       CC                   int3     
						;; bbWeight=0    PerfScore 0.00

; Total bytes of code 26, prolog size 4, PerfScore 7.60, (MethodHash=bdca8644) for method V128:GetUInt32(int):int:this
; ============================================================
```