@FICTURE7 2020-04-10
For some methods in `V128`:

Before
|    Method |     Mean |     Error |    StdDev |
|---------- |---------:|----------:|----------:|
|      Zero | 1.599 ns | 0.0012 ns | 0.0011 ns |
| GetUInt64 | 3.594 ns | 0.0138 ns | 0.0115 ns |
| GetUInt32 | 3.572 ns | 0.0046 ns | 0.0043 ns |

After
|    Method |      Mean |     Error |    StdDev |
|---------- |----------:|----------:|----------:|
|      Zero | 0.0000 ns | 0.0000 ns | 0.0000 ns |
| GetUInt64 | 0.3313 ns | 0.0015 ns | 0.0012 ns |
| GetUInt32 | 0.5286 ns | 0.0013 ns | 0.0012 ns |

~8.8X speed up.

For the software fallback of SHA256:

Before
|     Method |     Mean |   Error |  StdDev |
|----------- |---------:|--------:|--------:|
| HashChoose | 177.1 ns | 0.13 ns | 0.12 ns |
|  HashLower | 397.3 ns | 0.41 ns | 0.39 ns |

After
|     Method |     Mean |    Error |   StdDev |
|----------- |---------:|---------:|---------:|
| HashChoose | 49.00 ns | 0.038 ns | 0.032 ns |
|  HashLower | 93.54 ns | 0.041 ns | 0.039 ns |

~4X speed up.

For `MemoryManager.WriteVector128`:

Before
|         Method |      Mean |    Error |   StdDev |
|--------------- |----------:|---------:|---------:|
|   AlignedWrite |  35.27 ns | 0.129 ns | 0.120 ns |
| UnalignedWrite | 237.07 ns | 0.220 ns | 0.195 ns |

After
|         Method |      Mean |    Error |   StdDev |
|--------------- |----------:|---------:|---------:|
|   AlignedWrite |  32.12 ns | 0.673 ns | 0.720 ns |
| UnalignedWrite | 229.90 ns | 2.379 ns | 2.226 ns |

I guess the cost of the other calls dwarfs the now inlined `GetUInt64`. This benchmark will be moot anyways when #856 gets merged, but this may give an insight on the speed up the new memory manager might produce especially for unaligned writes.

This PR might also speed up atomic loads on `V128` (`ReadVector128Exclusive`), but I did not measure these.

<details>
<summary>Benchmark Details</summary>

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.720 (1909/November2018Update/19H2)
Intel Pentium CPU G2020 2.90GHz, 1 CPU, 2 logical and 2 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT


```

```csharp

public class V128Benchmark
{
    public int Index { get; set; } = 0;

    [Benchmark]
    public V128 Zero()
    {
        return V128.Zero;
    }

    // Used GetUInt64() & GetUInt32() when measuring master.

    [Benchmark]
    public ulong GetUInt64()
    {
        return default(V128).Get<ulong>(Index);
    }

    [Benchmark]
    public uint GetUInt32()
    {
        return default(V128).Get<uint>(Index);
    }
}

public class WriteV128Benchmark
{
    private const int RamSize = 4 * 1024;

    private IntPtr _ram;
    private MemoryManager _memory;

    public int AlignedPosition { get; set; } = 0x0;
    public int UnalignedPosition { get; set; } = 0x7;

    [GlobalSetup]
    public void Setup()
    {
        _ram = MemoryManagement.Allocate(RamSize);
        _memory = new MemoryManager(_ram);
        _memory.Map(0x0000, 0x0000, RamSize);
    }

    [GlobalCleanup]
    public void Cleanup()
    {
        _memory.Dispose();
    }

    [Benchmark]
    public void AlignedWrite()
    {
        _memory.WriteVector128(AlignedPosition, default);
    }

    [Benchmark]
    public void UnalignedWrite()
    {
        _memory.WriteVector128(UnalignedPosition, default);
    }
}

public class Sha256Benchmark
{
    [Benchmark]
    public V128 HashChoose()
    {
        return SoftFallback.HashChoose(default, 0, default);
    }

    [Benchmark]
    public V128 HashLower()
    {
        return SoftFallback.HashLower(default, default, default);
    }
}

```
</details>