@FICTURE7 2020-04-10
If this gets approved I'll try to add documentation, since it seems to be desired judging from the recent PRs.